+++ app/src/pages/Detailed/Detailed.js
@@ -18,6 +18,7 @@ import EthSpent from './../../pages/EthSpent'
 import { calculateBTCVolume, calculateBTCMarketcap } from './../../utils/utils'
 import { checkHasPremium, checkIsLoggedIn } from './../UserSelectors'
 import { millify } from './../../utils/formatting'
+import DetailedEthTopTransactions from './DetailedEthTopTransactions'
 import {
   projectBySlugGQL,
   TwitterDataGQL,
@@ -278,32 +279,7 @@ export const Detailed = ({
           project.isERC20 &&
           project.ethTopTransactions &&
           project.ethTopTransactions.length > 0 && (
-            <PanelBlock
-              isLoading={Project.loading}
-              title='Top ETH Transactions'
-            >
-              <div>
-                {project.ethTopTransactions &&
-                  project.ethTopTransactions.map((transaction, index) => (
-                    <div className='top-eth-transaction' key={index}>
-                      <div className='top-eth-transaction__hash'>
-                        <a
-                          href={`https://etherscan.io/tx/${
-                            transaction.trxHash
-                          }`}
-                        >
-                          {transaction.trxHash}
-                        </a>
-                      </div>
-                      <div>
-                        {millify(transaction.trxValue, 2)}
-                        &nbsp; | &nbsp;
-                        {moment(transaction.datetime).fromNow()}
-                      </div>
-                    </div>
-                  ))}
-              </div>
-            </PanelBlock>
+            <DetailedEthTopTransactions Project={Project} />
           )}
       </div>
     </div>